History is made lighter by storing only higher level calls

+++ trunk/@iData/isinf.m
@@ -8,11 +8,11 @@ function a = isinf(a)
 % output: b: object or array (iData)
 % ex:     b=isinf(a);
 %
-% Version: $Revision: 1.3 $
+% Version: $Revision: 1.4 $
 % See also iData, iData/sign, iData/isreal, iData/isfinite, iData/isnan,
 %          iData/isinf, iData/isfloat, iData/isinterger,
 %          iData/isnumeric, iData/islogical, iData/isscalar, 
 %          iData/isvector, iData/issparse
 
 a = iData_private_unary(a, 'isinf');
-a = double(a);
+a = uint8(a);
